Add feature for breaking cyclic dependences
See https://github.com/bazelbuild/rules_python/pull/1166

In our case, the particular dependency that has cycles and is
problematic is apache-airflow. This adds an `excluded_deps` keyword
argument to the `package_annotation` function, which is used to filter
out the deps in `excluded_deps` from the wheel for the dependency using
that annotation.

This lets us cut cycles, at the cost of having to manually specify
`requirement('dependency-that-we-cut-connection-to')` anywhere where we
were using the parent dep and expecting
`dependency-that-we-cut-connection-to` to get pulled in implicitly.